+++ src/_gettsim/transfers/arbeitsl_geld_2/kost_unterk.py
@@ -29,8 +29,8 @@ def arbeitsl_geld_2_kost_unterk_m_bg_bis_2022(
 
 @policy_info(start_date="2023-01-01", name_in_dag="arbeitsl_geld_2_kost_unterk_m_bg")
 def arbeitsl_geld_2_kost_unterk_m_bg_ab_2023(
-    bruttokaltmiete_m_hh: float,
-    heizkosten_m_hh: float,
+    bruttokaltmiete_m_bg: float,
+    heizkosten_m_bg: float,
     bürgerg_bezug_vorj: bool,
     _arbeitsl_geld_2_berechtigte_wohnfläche_bg: float,
     _arbeitsl_geld_2_warmmiete_pro_qm_m_bg: float,
@@ -43,10 +43,10 @@ def arbeitsl_geld_2_kost_unterk_m_bg_ab_2023(
 
     Parameters
     ----------
-    bruttokaltmiete_m_hh
-        See basic input variable :ref:`bruttokaltmiete_m_hh <bruttokaltmiete_m_hh>`.
-    heizkosten_m_hh
-        See basic input variable :ref:`heizkosten_m_hh <heizkosten_m_hh>`.
+    bruttokaltmiete_m_bg
+        See :func:`bruttokaltmiete_m_bg`.
+    heizkosten_m_bg
+        See :func:`heizkosten_m_bg`.
     bürgerg_bezug_vorj
         See basic input variable :ref:`bürgerg_bezug_vorj <bürgerg_bezug_vorj>`.
     _arbeitsl_geld_2_berechtigte_wohnfläche_bg
@@ -65,15 +65,15 @@ def arbeitsl_geld_2_kost_unterk_m_bg_ab_2023(
             * _arbeitsl_geld_2_warmmiete_pro_qm_m_bg
         )
     else:
-        out = bruttokaltmiete_m_hh + heizkosten_m_hh
+        out = bruttokaltmiete_m_bg + heizkosten_m_bg
 
     return out
 
 
 def _arbeitsl_geld_2_warmmiete_pro_qm_m_bg(
-    bruttokaltmiete_m_hh: float,
-    heizkosten_m_hh: float,
-    wohnfläche_hh: float,
+    bruttokaltmiete_m_bg: float,
+    heizkosten_m_bg: float,
+    wohnfläche_bg: float,
     arbeitsl_geld_2_params: dict,
 ) -> float:
     """Calculate rent per square meter.
@@ -82,19 +82,19 @@ def _arbeitsl_geld_2_warmmiete_pro_qm_m_bg(
 
     Parameters
     ----------
-    bruttokaltmiete_m_hh
-        See basic input variable :ref:`bruttokaltmiete_m_hh <bruttokaltmiete_m_hh>`.
-    heizkosten_m_hh
-        See basic input variable :ref:`heizkosten_m_hh <heizkosten_m_hh>`.
-    wohnfläche_hh
-        See basic input variable :ref:`wohnfläche_hh <wohnfläche_hh>`.
+    bruttokaltmiete_m_bg
+        See :func:`bruttokaltmiete_m_bg`.
+    heizkosten_m_bg
+        See :func:`heizkosten_m_bg`.
+    wohnfläche_bg
+        See function :func:`wohnfläche_bg`.
 
     Returns
     -------
     Integer with the total amount of rental costs per squaremeter.
 
     """
-    out = (bruttokaltmiete_m_hh + heizkosten_m_hh) / wohnfläche_hh
+    out = (bruttokaltmiete_m_bg + heizkosten_m_bg) / wohnfläche_bg
 
     # Consider maximum considered rent per square meter
     out = min(out, arbeitsl_geld_2_params["max_miete_pro_qm"]["max"])
@@ -103,9 +103,9 @@ def _arbeitsl_geld_2_warmmiete_pro_qm_m_bg(
 
 
 def _arbeitsl_geld_2_berechtigte_wohnfläche_bg(
-    wohnfläche_hh: float,
+    wohnfläche_bg: float,
     bewohnt_eigentum_hh: bool,
-    anz_personen_hh: int,
+    anz_personen_bg: int,
     arbeitsl_geld_2_params: dict,
 ) -> float:
     """Calculate size of dwelling eligible to claim.
@@ -114,12 +114,12 @@ def _arbeitsl_geld_2_berechtigte_wohnfläche_bg(
 
     Parameters
     ----------
-    wohnfläche_hh
-        See basic input variable :ref:`wohnfläche_hh <wohnfläche_hh>`.
+    wohnfläche_bg
+        See function :func:`wohnfläche_bg`.
     bewohnt_eigentum_hh
         See basic input variable :ref:`bewohnt_eigentum_hh <bewohnt_eigentum_hh>`.
-    anz_personen_hh
-        See :func:`anz_personen_hh`.
+    anz_personen_bg
+        See :func:`anz_personen_bg`.
 
     Returns
     -------
@@ -127,24 +127,90 @@ def _arbeitsl_geld_2_berechtigte_wohnfläche_bg(
 
     """
 
-    # TODO(@MImmesberger): Use variables on the bg level instead of hh level here.
-    # https://github.com/iza-institute-of-labor-economics/gettsim/issues/710
     params = arbeitsl_geld_2_params["berechtigte_wohnfläche_eigentum"]
     max_anzahl_direkt = params["max_anzahl_direkt"]
     if bewohnt_eigentum_hh:
-        if anz_personen_hh <= max_anzahl_direkt:
-            maximum = params[anz_personen_hh]
+        if anz_personen_bg <= max_anzahl_direkt:
+            maximum = params[anz_personen_bg]
         else:
             maximum = (
                 params[max_anzahl_direkt]
-                + (anz_personen_hh - max_anzahl_direkt) * params["je_weitere_person"]
+                + (anz_personen_bg - max_anzahl_direkt) * params["je_weitere_person"]
             )
     else:
         maximum = (
             arbeitsl_geld_2_params["berechtigte_wohnfläche_miete"]["single"]
-            + max(anz_personen_hh - 1, 0)
+            + max(anz_personen_bg - 1, 0)
             * arbeitsl_geld_2_params["berechtigte_wohnfläche_miete"][
                 "je_weitere_person"
             ]
         )
-    return min(wohnfläche_hh, maximum)
+    return min(wohnfläche_bg, maximum)
+
+
+def bruttokaltmiete_m_bg(
+    bruttokaltmiete_m_hh: float,
+    _anteil_personen_in_haushalt_bg: float,
+) -> float:
+    """Share of household's monthly rent attributed to the Bedarfsgemeinschaft.
+
+    Reference:
+    BSG Urteil v. 09.03.2016 - B 14 KG 1/15 R.
+    BSG Urteil vom 15.04.2008 - B 14/7b AS 58/06 R.
+
+    Parameters
+    ----------
+    bruttokaltmiete_m_hh
+        See basic input variable :ref:`bruttokaltmiete_m_hh <bruttokaltmiete_m_hh>`.
+    _anteil_personen_in_haushalt_bg
+        See :func:`_anteil_personen_in_haushalt_bg`.
+
+    Returns
+    -------
+
+    """
+    return bruttokaltmiete_m_hh * _anteil_personen_in_haushalt_bg
+
+
+def heizkosten_m_bg(
+    heizkosten_m_hh: float,
+    _anteil_personen_in_haushalt_bg: float,
+) -> float:
+    """Share of household's heating expenses attributed to the Bedarfsgemeinschaft.
+
+    Reference:
+    BSG Urteil v. 09.03.2016 - B 14 KG 1/15 R.
+    BSG Urteil vom 15.04.2008 - B 14/7b AS 58/06 R.
+
+    Parameters
+    ----------
+    heizkosten_m_hh
+        See basic input variable :ref:`heizkosten_m_hh <heizkosten_m_hh>`.
+    _anteil_personen_in_haushalt_bg
+        See :func:`_anteil_personen_in_haushalt_bg`.
+
+    Returns
+    -------
+
+    """
+    return heizkosten_m_hh * _anteil_personen_in_haushalt_bg
+
+
+def wohnfläche_bg(
+    wohnfläche_hh: float,
+    _anteil_personen_in_haushalt_bg: float,
+) -> float:
+    """Share of household's dwelling size attributed to the Bedarfsgemeinschaft.
+
+    Parameters
+    ----------
+    wohnfläche_hh
+        See basic input variable :ref:`wohnfläche_hh <wohnfläche_hh>`.
+    _anteil_personen_in_haushalt_bg
+        See :func:`_anteil_personen_in_haushalt_bg`.
+
+    Returns
+    -------
+
+    """
+    return wohnfläche_hh * _anteil_personen_in_haushalt_bg